requirements.txt
absl-py==1.4.0
astor==0.8.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
cached-property==1.5.2
certifi==2021.5.30
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
dataclasses==0.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
gast==0.5.4
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
importlib-metadata==4.8.3
imutils==0.5.4
ipykernel @ file:///D:/bld/ipykernel_1620913140366/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1609697945785/work
ipython-genutils==0.2.0
jedi @ file:///D:/bld/jedi_1605054785407/work
jieba==0.42.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1642858610849/work
jupyter-core @ file:///D:/bld/jupyter_core_1631852848502/work
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.19.5
pandas==1.1.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1595548966091/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==3.19.6
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
pywin32==301
pyzmq @ file:///D:/bld/pyzmq_1631793604214/work
scipy==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado @ file:///D:/bld/tornado_1610094881553/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1631041982274/work
typing_extensions==4.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.15.0
zipp==3.6.0
